+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - Added: wildcard support for parameter title in Get-IdoitObjectByFilter
 - Fixed: Error passing Id parameter to Get-IdoItCategory
+
+##v0.1.0.42 - 01/15/2018
+
+- Added: Create a constant cache in the users romaing profile when calling Connect-IdoIt
+- Added: DynamicParameter validate set for parameter <Category> that pulls constants from the psidoit cache